Bullets fixed, Levels soon
According to the youtube vid I found ( https://youtu.be/c2mkyW_TymY ), I'm unsure what I'm missing, but I cannot get my player to go to level 2. That and I fear that I must make a level to send you directly to the win scene. However, all backgrounds should be fixed alongside this. The lag may still be apparent after 30 seconds from shooting a bullet (which will definitely need fixing). If other levels do not work, perhaps we can implement a 'universal timer', as loading into "a new level" resets the timer as well. This may become a universal thing regardless of if its per level or not.